feat(transcript): add useTranscript hook and enhance transcript view
The changes in this commit introduce a new `useTranscript` hook and enhance the transcript view in the desktop application. The key changes are:

1. Implement the `useTranscript` hook to manage the state and logic related to the transcript, including fetching the timeline data, handling live updates, and managing the selected language.
2. Refactor the `TranscriptBase` component to use the new `useTranscript` hook and provide a more modular and reusable implementation.
3. Remove the legacy `TranscriptBase` component and its associated logic, as it has been replaced by the new implementation.

These changes aim to improve the overall quality and maintainability of the transcript feature in the desktop application.
