+++ apps/desktop/src/components/right-panel/hooks/useTranscript.ts
