+++ apps/desktop/src/components/right-panel/hooks/useTranscriptWidget.ts
