--- apps/desktop/src/components/right-panel/views/legacy/index.tsx
@@ -1,14 +1,0 @@
-import { WidgetFullSize, WidgetFullSizeWrapper } from "./ui";
-
-import { TranscriptBase } from "./base";
-
-const TranscriptFull: WidgetFullSize = ({ queryClient, onMinimize }) => {
-  return (
-    <TranscriptBase
-      WrapperComponent={WidgetFullSizeWrapper}
-      wrapperProps={{ queryClient, onMinimize }}
-    />
-  );
-};
-
-export default TranscriptFull;
